+++ src/sumset.c
@@ -33,9 +33,6 @@
 #include "util.h"
 #include "trace.h"
 
-LIBRSYNC_EXPORT const int RS_MD4_SUM_LENGTH = 16;
-LIBRSYNC_EXPORT const int RS_BLAKE2_SUM_LENGTH = 32;
-
 static void rs_block_sig_init(rs_block_sig_t *sig, rs_weak_sum_t weak_sum,
                               rs_strong_sum_t *strong_sum, int strong_len)
 {
@@ -84,6 +81,9 @@ static inline int rs_block_match_cmp(rs_block_match_t *match,
                   match->signature->strong_sum_len);
 }
 
+/* Disable mix32() in the hashtable because RabinKarp doesn't need it. We
+   manually apply mix32() to rollsums before using them in the hashtable. */
+#define HASHTABLE_NMIX32
 /* Instantiate hashtable for rs_block_sig and rs_block_match. */
 #define ENTRY rs_block_sig
 #define MATCH rs_block_match
@@ -124,12 +124,14 @@ rs_result rs_signature_init(rs_signature_t *sig, int magic, int block_len,
     int max_strong_len;
 
     /* Check and set default arguments. */
-    magic = magic ? magic : RS_BLAKE2_SIG_MAGIC;
+    magic = magic ? magic : RS_RK_BLAKE2_SIG_MAGIC;
     switch (magic) {
     case RS_BLAKE2_SIG_MAGIC:
+    case RS_RK_BLAKE2_SIG_MAGIC:
         max_strong_len = RS_BLAKE2_SUM_LENGTH;
         break;
     case RS_MD4_SIG_MAGIC:
+    case RS_RK_MD4_SIG_MAGIC:
         max_strong_len = RS_MD4_SUM_LENGTH;
         break;
     default:
@@ -176,6 +178,9 @@ rs_block_sig_t *rs_signature_add_block(rs_signature_t *sig,
                                        rs_strong_sum_t *strong_sum)
 {
     rs_signature_check(sig);
+    /* Apply mix32() to rollsum weaksums to improve their distribution. */
+    if (rs_signature_weaksum_kind(sig) == RS_ROLLSUM)
+        weak_sum = mix32(weak_sum);
     /* If block_sigs is full, allocate more space. */
     if (sig->count == sig->size) {
         sig->size = sig->size ? sig->size * 2 : 16;